refactor: extract useClickPlacement (handleClick pick/place dispatcher) from SketchCanvas
