+++ src/components/canvas/SketchCanvas.tsx
@@ -27,7 +27,6 @@ import { filletFeatureFromPoint, filletFeatureFromRadius, filletRadiusFromPoint,
 import {
   buildPendingDraftProfile,
   buildPendingProfile,
-  compositeDraftPoints,
   drawCompositeDraft,
   drawSnapIndicator,
   findOpenProfileExtensionEndpoint,
@@ -58,10 +57,11 @@ import { useOffsetWorkflow } from './useOffsetWorkflow'
 import { useTransformExactWorkflow } from './useTransformExactWorkflow'
 import { useCreationWorkflow } from './useCreationWorkflow'
 import { useCanvasKeyboard } from './useCanvasKeyboard'
+import { useClickPlacement } from './useClickPlacement'
 import { useSnapPreview } from './useSnapPreview'
 import { useCanvasContextMenu } from './useCanvasContextMenu'
 import { drawDimensions, drawPendingDimensionPreview, drawTapeMeasure, pickDimensionAt } from './dimensionRendering'
-import { circleEdgeAnchorFromPoint, offsetForCursor } from '../../sketch/dimensions'
+import { offsetForCursor } from '../../sketch/dimensions'
 import {
   drawFeature,
   drawMoveGuide,
@@ -99,7 +99,6 @@ import {
   drawSketchEditPreviewPoint,
   drawStockOutline,
   drawTabFootprint,
-  hitBackdrop,
 } from './scenePrimitives'
 import { generateTextShapes } from '../../text'
 import {
@@ -110,7 +109,7 @@ import {
   profileVertices,
   rectProfile,
 } from '../../types/project'
-import type { Clamp, DimensionAnchor, DimensionAnnotation, OperationKind, Point, SketchFeature, Tab } from '../../types/project'
+import type { Clamp, OperationKind, Point, SketchFeature, Tab } from '../../types/project'
 import { compatibleFeatureIdsForOperation } from '../cam/operationValidity'
 import { formatLength, parseLengthInput } from '../../utils/units'
 import { useAxisLock, lockModeGuideColor } from '../../sketch/useAxisLock'
@@ -2958,505 +2957,6 @@ export const SketchCanvas = forwardRef<SketchCanvasHandle, SketchCanvasProps>(fu
     }
   }
 
-  function handleClick(event: MouseEvent<HTMLCanvasElement>) {
-    if (suppressClickRef.current) {
-      suppressClickRef.current = false
-      return
-    }
-
-    if (zoomWindowActive) {
-      return
-    }
-
-    if (didPanRef.current) {
-      didPanRef.current = false
-      return
-    }
-
-    const selection = selectionRef.current
-    const project = projectRef.current
-    const pendingAdd = pendingAddRef.current
-    const pendingMove = pendingMoveRef.current
-    const pendingTransform = pendingTransformRef.current
-    const pendingOffset = pendingOffsetRef.current
-    const pendingShapeAction = pendingShapeActionRef.current
-    const viewState = viewStateRef.current
-    const dimensionEdit = dimEdit.dimensionEditRef.current
-    if (isDraggingNodeRef.current) return
-
-    const point = canvasCoordinates(event)
-    const canvas = canvasRef.current
-    if (!canvas) return
-
-    const vt = computeViewTransform(project.stock, canvas.width, canvas.height, viewState)
-    const world = canvasToWorld(point.cx, point.cy, vt)
-    const pendingConstraint = pendingConstraintRef.current
-    const constraintRefPickingClick = !!pendingConstraint && !!pendingConstraint.anchor && !pendingConstraint.reference
-    const resolvedSnap = snap.resolveCurrentSketchSnap(world, vt, {
-      excludeActiveEditGeometry: constraintRefPickingClick,
-    })
-    const pickedPoint = snap.requiresResolvedSnapForPointPick() && !resolvedSnap.mode ? null : resolvedSnap.point
-
-    // ── Delete-dimension mode: click a dimension to remove it (stays armed) ──
-    if (!pendingAdd && dimensionDeleteArmedRef.current) {
-      if (project.meta.showDimensions) {
-        const hit = pickDimensionAt(project, vt, point, 8)
-        if (hit) {
-          deleteDimensionAnnotation(hit)
-          deleteHoverDimIdRef.current = null
-        }
-      }
-      return
-    }
-
-    // ── Tape measure: each click sets/advances the transient measurement ──
-    if (!pendingAdd && tapeMeasureRef.current) {
-      tapeMeasureClick(resolvedSnap.point)
-      return
-    }
-
-    // ── Permanent dimension placement ──
-    const pendingDim = pendingAdd ? null : pendingDimensionRef.current
-    if (pendingDim) {
-      const anchor: DimensionAnchor = resolvedSnap.anchor ?? { kind: 'free', point: resolvedSnap.point }
-      const need = pendingDim.type === 'angle' ? 3 : 2
-      const picked = [pendingDim.a, pendingDim.b, pendingDim.c].filter(Boolean).length
-      // Radius/diameter have no cursor-driven offset (line is always center→edge),
-      // so commit immediately on the final anchor click instead of asking for an
-      // extra "click to place" step.
-      const isRadial = pendingDim.type === 'radius' || pendingDim.type === 'diameter'
-      // For radius/diameter the first click MUST identify a circle/arc center —
-      // otherwise the dimension would be measured between arbitrary points and
-      // the value would be meaningless. Silently ignore non-center clicks; the
-      // CanvasWorkflowPanel already says "Click the circle / arc center".
-      if (isRadial && picked === 0 && anchor.kind !== 'center') {
-        return
-      }
-      if (isRadial && picked === need - 1) {
-        // If the edge click landed off-circle (or via a non-anchored snap), but
-        // anchor a pins the circle's center, project onto that circle and store
-        // an angle-relative anchor so the dim direction follows the feature.
-        const edgeAnchor =
-          anchor.kind === 'free' && pendingDim.a
-            ? (circleEdgeAnchorFromPoint(anchor.point, pendingDim.a, project) ?? anchor)
-            : anchor
-        addDimensionAnnotation({
-          type: pendingDim.type,
-          a: pendingDim.a!,
-          b: edgeAnchor,
-          offset: 0,
-          visible: true,
-          locked: false,
-          textOverride: null,
-          precisionOverride: null,
-        })
-        cancelPendingDimension()
-        return
-      }
-      if (picked < need) {
-        pendingDimensionPick(anchor)
-        return
-      }
-      // All anchors picked → this click chooses the offset and commits.
-      const temp: DimensionAnnotation = {
-        id: '__commit__',
-        type: pendingDim.type,
-        a: pendingDim.a!,
-        b: pendingDim.b ?? undefined,
-        c: pendingDim.c ?? undefined,
-        offset: 0,
-        visible: true,
-        locked: false,
-        textOverride: null,
-        precisionOverride: null,
-      }
-      const off = offsetForCursor(temp, project, world) ?? 0
-      addDimensionAnnotation({
-        type: pendingDim.type,
-        a: pendingDim.a!,
-        b: pendingDim.b ?? undefined,
-        c: pendingDim.c ?? undefined,
-        offset: off,
-        visible: true,
-        locked: false,
-        textOverride: null,
-        precisionOverride: null,
-      })
-      cancelPendingDimension()
-      return
-    }
-
-    // ── Select an existing dimension annotation (plain select mode) ──
-    if (
-      selection.mode === 'feature'
-      && !pendingAdd && !pendingMove && !pendingTransform && !pendingOffset
-      && !pendingShapeAction && !pendingConstraint
-      && project.meta.showDimensions
-    ) {
-      const hitDim = pickDimensionAt(project, vt, point, 8)
-      if (hitDim) {
-        selectAnnotation(hitDim)
-        return
-      }
-      if (selectedAnnotationIdRef.current) {
-        selectAnnotation(null)
-      }
-    }
-
-    if (pendingConstraint && !pendingConstraint.anchor) {
-      if (!pickedPoint) {
-        return
-      }
-      setConstraintAnchor({
-        point: pickedPoint,
-        snapMode: resolvedSnap.mode,
-      })
-      return
-    }
-
-    if (pendingConstraint && pendingConstraint.anchor && !pendingConstraint.reference) {
-      if (!pickedPoint) {
-        return
-      }
-      const lockedPickedPoint = applyLock(pickedPoint, pendingConstraint.anchor.point)
-      const targetFeatureId = findHitFeatureId(lockedPickedPoint, project.features, vt)
-      const targetId =
-        targetFeatureId && targetFeatureId !== pendingConstraint.featureId
-          ? targetFeatureId
-          : null
-      setConstraintReference({
-        point: lockedPickedPoint,
-        featureId: targetId,
-        snapMode: resolvedSnap.mode,
-        segment: resolvedSnap.mode === 'perpendicular' ? resolvedSnap.perpendicularSegment : undefined,
-      })
-      const dx = pendingConstraint.anchor.point.x - lockedPickedPoint.x
-      const dy = pendingConstraint.anchor.point.y - lockedPickedPoint.y
-      const currentDistance = Math.hypot(dx, dy)
-      constraint.setConstraintDistanceInput(formatLength(currentDistance, project.meta.units))
-      return
-    }
-
-    if (selection.mode === 'sketch_edit') {
-      if (selection.selectedNode?.type === 'feature' && selection.selectedFeatureId) {
-        const feature = editableFeature()
-        if (feature && selection.sketchEditTool === 'add_point') {
-          if (pendingSketchExtensionRef.current) {
-            const sourceEndpoint = endpointFromSketchExtension(pendingSketchExtensionRef.current.kind)
-            const targetEndpoint = findOpenEndpointHit(world, vt, {
-              exclude: {
-                featureId: selection.selectedFeatureId,
-                endpoint: sourceEndpoint,
-                anchor: pendingSketchExtensionRef.current.anchor,
-              },
-            })
-            if (targetEndpoint) {
-              const joined = joinOpenFeatureEndpoints(
-                selection.selectedFeatureId,
-                sourceEndpoint,
-                targetEndpoint.featureId,
-                targetEndpoint.endpoint,
-              )
-              if (joined) {
-                pendingSketchExtensionRef.current = null
-                sketchEditPreviewRef.current = null
-                scheduleDraw()
-              }
-              return
-            }
-
-            if (!pickedPoint) {
-              return
-            }
-            const lockedPickedPoint = applyLock(pickedPoint, pendingSketchExtensionRef.current.anchor)
-            insertFeaturePoint(selection.selectedFeatureId, {
-              kind: pendingSketchExtensionRef.current.kind,
-              point: lockedPickedPoint,
-            })
-            pendingSketchExtensionRef.current = null
-            sketchEditPreviewRef.current = null
-            scheduleDraw()
-            return
-          }
-
-          const endpoint = findOpenProfileExtensionEndpoint(feature.sketch.profile, world, vt)
-          if (endpoint) {
-            pendingSketchExtensionRef.current = endpoint
-            sketchEditPreviewRef.current = { point: endpoint.anchor, mode: 'add_point' }
-            scheduleDraw()
-            return
-          }
-
-          if (!pickedPoint) {
-            return
-          }
-
-          const target = findSketchInsertTarget(feature.sketch.profile, pickedPoint, vt)
-          if (target?.kind === 'segment') {
-            insertFeaturePoint(selection.selectedFeatureId, target)
-          }
-          return
-        }
-
-        if (feature && selection.sketchEditTool === 'delete_point') {
-          const control = hitEditableControl(point, { includeSegments: false })
-          if (control?.kind === 'anchor') {
-            deleteFeaturePoint(selection.selectedFeatureId, control.index)
-          }
-          return
-        }
-
-        if (feature && selection.sketchEditTool === 'delete_segment') {
-          const target = findSketchSegmentHit(feature.sketch.profile, world, vt)
-          if (target) {
-            deleteFeatureSegment(selection.selectedFeatureId, target.segmentIndex)
-            sketchEditPreviewRef.current = null
-            scheduleDraw()
-          }
-          return
-        }
-
-        if (feature && selection.sketchEditTool === 'disconnect') {
-          const control = hitEditableControl(point, { includeSegments: false })
-          if (control?.kind === 'anchor') {
-            disconnectFeaturePoint(selection.selectedFeatureId, control.index)
-            sketchEditPreviewRef.current = null
-            scheduleDraw()
-          }
-          return
-        }
-
-        if (feature && selection.sketchEditTool === 'fillet') {
-          if (pendingSketchFilletRef.current) {
-            const typedRadius = fillet.filletDimensionEditRef.current
-              ? parseLengthInput(fillet.filletDimensionEditRef.current.radius, project.meta.units)
-              : null
-            if (typedRadius !== null && typedRadius > 0) {
-              filletFeaturePoint(selection.selectedFeatureId, pendingSketchFilletRef.current.anchorIndex, typedRadius)
-            } else {
-              if (!pickedPoint) {
-                return
-              }
-              const radius = filletRadiusFromPoint(feature, pendingSketchFilletRef.current.anchorIndex, pickedPoint)
-              if (radius) {
-                filletFeaturePoint(selection.selectedFeatureId, pendingSketchFilletRef.current.anchorIndex, radius)
-              }
-            }
-            pendingSketchFilletRef.current = null
-            sketchEditPreviewRef.current = null
-            fillet.setFilletDimensionEdit(null)
-            scheduleDraw()
-            return
-          }
-
-          const control = hitEditableControl(point, { includeSegments: false })
-          if (control?.kind === 'anchor') {
-            pendingSketchFilletRef.current = {
-              anchorIndex: control.index,
-              corner: anchorPointForIndex(feature.sketch.profile, control.index),
-            }
-            sketchEditPreviewRef.current = { point: pendingSketchFilletRef.current.corner, mode: 'add_point' }
-            scheduleDraw()
-          }
-          return
-        }
-      }
-
-      if (selection.selectedNode?.type === 'feature' && selection.selectedFeatureId && !selection.sketchEditTool && !dimensionEdit) {
-        const feature = editableFeature()
-        if (feature) {
-          const control = hitEditableControl(point)
-          if (control && (control.kind === 'segment' || control.kind === 'anchor' || control.kind === 'arc_handle')) {
-            const steps = dimEdit.computeEditStepsForControl(feature.sketch.profile, control)
-            if (steps.length > 0) {
-              dimEdit.editDimStepsRef.current = steps
-              dimEdit.editDimStepIndexRef.current = 0
-              dimEdit.dimensionEditFeatureIdRef.current = selection.selectedFeatureId
-              beginHistoryTransaction()
-              dimEdit.applyEditDimStep(0, steps, selection.selectedFeatureId, project.meta.units)
-              return
-            }
-          }
-        }
-      }
-
-      return
-    }
-
-    if (dimensionEdit) {
-      dimEdit.commitEditDimension()
-      return
-    }
-
-    if (pendingAdd) {
-      if (!pickedPoint) {
-        return
-      }
-
-      const snapped = pickedPoint
-
-      if (pendingAdd.shape === 'origin') {
-        originPreviewPointRef.current = null
-        placeOriginAt(snapped)
-        setPendingPreviewPointRef(null)
-        return
-      }
-
-      if (pendingAdd.shape === 'polygon' || pendingAdd.shape === 'spline') {
-        const lastPoint = pendingAdd.points[pendingAdd.points.length - 1]
-        if (pendingAdd.points.length >= 3 && isLoopCloseCandidate(point, pendingAdd.points, vt, POLYGON_CLOSE_RADIUS)) {
-          completePendingPolygon()
-          setPendingPreviewPointRef(null)
-          return
-        }
-        const lockedSnapped = lastPoint ? applyLock(snapped, lastPoint) : snapped
-        if (!lastPoint || lastPoint.x !== lockedSnapped.x || lastPoint.y !== lockedSnapped.y) {
-          addPendingPolygonPoint(lockedSnapped)
-        }
-        setPendingPreviewPointRef({ point: lockedSnapped, session: pendingAdd.session })
-      } else if ((pendingAdd.shape === 'rect' || pendingAdd.shape === 'circle' || pendingAdd.shape === 'ellipse' || pendingAdd.shape === 'tab' || pendingAdd.shape === 'clamp') && !pendingAdd.anchor) {
-        setPendingAddAnchor(snapped)
-        setPendingPreviewPointRef({ point: snapped, session: pendingAdd.session })
-      } else if (pendingAdd.shape === 'rect' || pendingAdd.shape === 'circle' || pendingAdd.shape === 'ellipse' || pendingAdd.shape === 'tab' || pendingAdd.shape === 'clamp') {
-        placePendingAddAt(snapped)
-        setPendingPreviewPointRef(null)
-      } else if (pendingAdd.shape === 'text') {
-        placePendingTextAt(snapped)
-        setPendingPreviewPointRef(null)
-      } else if (pendingAdd.shape === 'composite') {
-        const draftPoints = compositeDraftPoints(pendingAdd)
-        const closeCandidate =
-          pendingAdd.currentMode !== 'arc' &&
-          !pendingAdd.pendingArcEnd &&
-          draftPoints.length >= 3 &&
-          isLoopCloseCandidate(point, draftPoints, vt, POLYGON_CLOSE_RADIUS)
-
-        if (closeCandidate) {
-          completePendingComposite()
-          setPendingPreviewPointRef(null)
-          return
-        }
-
-        const compositeOrigin = pendingAdd.pendingArcEnd ?? pendingAdd.lastPoint ?? pendingAdd.start
-        const lockedCompositeSnapped = compositeOrigin ? applyLock(snapped, compositeOrigin) : snapped
-        addPendingCompositePoint(lockedCompositeSnapped)
-        setPendingPreviewPointRef({ point: lockedCompositeSnapped, session: pendingAdd.session })
-      }
-      return
-    }
-
-    if (pendingMove) {
-      if (!pickedPoint) {
-        return
-      }
-
-      const snapped = pickedPoint
-
-      if (!pendingMove.fromPoint) {
-        setPendingMoveFrom(snapped)
-        setPendingMovePreviewPointRef({ point: snapped, session: pendingMove.session })
-      } else if (!pendingMove.toPoint) {
-        const lockedSnapped = applyLock(snapped, pendingMove.fromPoint)
-        move.beginMoveDistanceEntry(lockedSnapped)
-      }
-      return
-    }
-
-    if (pendingTransform) {
-      if (!pickedPoint) {
-        return
-      }
-
-      const snapped = pickedPoint
-      const constrainedPoint =
-        pendingTransform.mode === 'resize' && pendingTransform.referenceStart && pendingTransform.referenceEnd
-          ? projectPointOntoLine(snapped, pendingTransform.referenceStart, pendingTransform.referenceEnd)
-          : snapped
-
-      if (!pendingTransform.referenceStart) {
-        setPendingTransformReferenceStart(snapped)
-        setPendingTransformPreviewPointRef({ point: snapped, session: pendingTransform.session })
-      } else if (!pendingTransform.referenceEnd) {
-        setPendingTransformReferenceEnd(snapped)
-        setPendingTransformPreviewPointRef({ point: snapped, session: pendingTransform.session })
-        if (pendingTransform.mode === 'mirror') {
-          completePendingTransform(snapped)
-          setPendingTransformPreviewPointRef(null)
-        }
-      } else if (pendingTransform.mode === 'rotate' && pendingTransform.keepOriginals) {
-        transformExact.setPendingRotateCopyPoint(constrainedPoint)
-      } else {
-        completePendingTransform(constrainedPoint)
-        setPendingTransformPreviewPointRef(null)
-      }
-      return
-    }
-
-    if (pendingOffset) {
-      const sourceFeatures = pendingOffset.entityIds
-        .map((featureId) => project.features.find((feature) => feature.id === featureId) ?? null)
-        .filter((feature): feature is SketchFeature => feature !== null)
-        .filter((feature) => feature.sketch.profile.closed)
-      if (!pickedPoint) {
-        return
-      }
-      const previewInput = resolveOffsetPreview(sourceFeatures, world, pickedPoint, resolvedSnap.mode, vt)
-      if (previewInput) {
-        completePendingOffset(previewInput.signedDistance)
-      } else {
-        cancelPendingOffset()
-      }
-      setPendingOffsetPreviewPointRef(null)
-      setPendingOffsetRawPreviewPointRef(null)
-      return
-    }
-
-    // Hit-test constraint labels for click-to-edit
-    if (!pendingConstraint && !pendingAdd && !pendingMove && !pendingTransform && !pendingOffset) {
-      for (const rect of constraintLabelRectsRef.current) {
-        if (
-          point.cx >= rect.cx - rect.halfW && point.cx <= rect.cx + rect.halfW &&
-          point.cy >= rect.cy - rect.halfH && point.cy <= rect.cy + rect.halfH
-        ) {
-          const feature = project.features.find((f) => f.id === rect.featureId)
-          const foundConstraint = feature?.sketch.constraints.find((c) => c.id === rect.constraintId)
-          if (foundConstraint && typeof foundConstraint.value === 'number') {
-            constraint.setConstraintEdit({
-              featureId: rect.featureId,
-              constraintId: rect.constraintId,
-              value: formatLength(foundConstraint.value, project.meta.units),
-              cx: rect.cx,
-              cy: rect.cy,
-            })
-          }
-          return
-        }
-      }
-    }
-
-    const hitClampId = findHitClampId(world, project.clamps)
-    if (hitClampId) {
-      selectClamp(hitClampId)
-      return
-    }
-
-    const hitTabId = findHitTabId(world, project.tabs)
-    if (hitTabId) {
-      selectTab(hitTabId)
-      return
-    }
-
-    const hitId = findHitFeatureId(world, project.features, vt)
-    const additive = event.metaKey || event.ctrlKey || event.shiftKey || multiSelectMode || !!pendingShapeAction
-    if (hitId) {
-      selectFeature(hitId, additive)
-    } else if (project.backdrop?.visible && hitBackdrop(world, project.backdrop)) {
-      selectBackdrop()
-    } else if (!additive) {
-      selectFeature(null)
-    }
-  }
-
   function handleWheelEvent(event: Pick<globalThis.WheelEvent, 'clientX' | 'clientY' | 'deltaY' | 'preventDefault'>) {
     if (zoomWindowActive) {
       return
@@ -3729,6 +3229,86 @@ export const SketchCanvas = forwardRef<SketchCanvasHandle, SketchCanvasProps>(fu
     resetLock,
   })
 
+  const clickPlacement = useClickPlacement({
+    suppressClickRef,
+    didPanRef,
+    isDraggingNodeRef,
+    zoomWindowActive,
+    multiSelectMode,
+    selectionRef,
+    projectRef,
+    pendingAddRef,
+    pendingMoveRef,
+    pendingTransformRef,
+    pendingOffsetRef,
+    pendingShapeActionRef,
+    viewStateRef,
+    pendingConstraintRef,
+    pendingDimensionRef,
+    dimensionDeleteArmedRef,
+    deleteHoverDimIdRef,
+    selectedAnnotationIdRef,
+    pendingSketchExtensionRef,
+    pendingSketchFilletRef,
+    sketchEditPreviewRef,
+    originPreviewPointRef,
+    tapeMeasureRef,
+    constraintLabelRectsRef,
+    canvasRef,
+    snap,
+    dimEdit,
+    move,
+    transformExact,
+    fillet,
+    constraint,
+    canvasCoordinates,
+    editableFeature,
+    endpointFromSketchExtension,
+    findOpenEndpointHit,
+    findSketchSegmentHit,
+    hitEditableControl,
+    scheduleDraw,
+    applyLock,
+    setPendingPreviewPointRef,
+    setPendingMovePreviewPointRef,
+    setPendingTransformPreviewPointRef,
+    setPendingOffsetPreviewPointRef,
+    setPendingOffsetRawPreviewPointRef,
+    tapeMeasureClick,
+    cancelPendingDimension,
+    addDimensionAnnotation,
+    pendingDimensionPick,
+    deleteDimensionAnnotation,
+    selectAnnotation,
+    selectFeature,
+    selectTab,
+    selectClamp,
+    selectBackdrop,
+    setConstraintAnchor,
+    setConstraintReference,
+    insertFeaturePoint,
+    joinOpenFeatureEndpoints,
+    deleteFeaturePoint,
+    deleteFeatureSegment,
+    disconnectFeaturePoint,
+    filletFeaturePoint,
+    setPendingAddAnchor,
+    placePendingAddAt,
+    placePendingTextAt,
+    placeOriginAt,
+    addPendingPolygonPoint,
+    completePendingPolygon,
+    addPendingCompositePoint,
+    completePendingComposite,
+    setPendingMoveFrom,
+    setPendingTransformReferenceStart,
+    setPendingTransformReferenceEnd,
+    completePendingTransform,
+    completePendingOffset,
+    cancelPendingOffset,
+    beginHistoryTransaction,
+  })
+
   const editingFeature =
     selection.mode === 'sketch_edit' && selection.selectedFeatureId
       ? project.features.find((feature) => feature.id === selection.selectedFeatureId) ??
@@ -3780,7 +3360,7 @@ export const SketchCanvas = forwardRef<SketchCanvasHandle, SketchCanvasProps>(fu
         onPointerUp={handlePointerUp}
         onPointerCancel={handlePointerUp}
         onPointerLeave={handlePointerLeave}
-        onClick={handleClick}
+        onClick={clickPlacement.handleClick}
         onDoubleClick={handleDoubleClick}
         onKeyDown={keyboard.handleKeyDown}
         onContextMenu={contextMenu.handleContextMenu}
